Add App.tsx in root directory and remove unnecessary components
