+++ App.tsx
@@ -1,0 +1,5 @@
+import RootLayout from './navigation/RootLayout';
+
+export default function Root() {
+  return <RootLayout />;
+}
